% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{format_columns}
\alias{format_columns}
\title{Format columns to be consistent within all databases}
\usage{
format_columns(db)
}
\value{
A dataframe with column names lower case and dotted.
}
\description{
Format columns to be consistent within all databases
}
