% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{get_database_names}
\alias{get_database_names}
\title{Get the names of the accessible databases}
\usage{
get_database_names()
}
\value{
a list of database names
}
\description{
Get the names of the accessible databases
}
