% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{get_lookup_by_name}
\alias{get_lookup_by_name}
\title{Get lookup table from name}
\usage{
get_lookup_by_name(lookup_name)
}
\value{
A dataframe containing lookup table.
}
\description{
Get lookup table from name
}
