% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{get_lookup_names}
\alias{get_lookup_names}
\title{Get the name of the available lookup tables}
\usage{
get_lookup_names()
}
\value{
A column of lookup table names.
}
\description{
Get the name of the available lookup tables
}
