% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{get_program_name_db}
\alias{get_program_name_db}
\title{Get UWL program names database}
\usage{
get_program_name_db()
}
\value{
A dataframe containing program name database.
}
\description{
Get UWL program names database
}
