% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{pre_professional_track_alias_lookup}
\alias{pre_professional_track_alias_lookup}
\title{General Eduation Category lookup table}
\usage{
pre_professional_track_alias_lookup
}
\description{
General Eduation Category lookup table
}
