% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{general_education_category_lookup}
\alias{general_education_category_lookup}
\title{General Education Category lookup table}
\usage{
general_education_category_lookup
}
\description{
General Education Category lookup table
}
