% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{get_student_program_db}
\alias{get_student_program_db}
\title{Get UWL student program data
currently from WINGS report UWL_SR_REQ_TERM_BY_COLLEGE}
\usage{
get_student_program_db(filename = NA)
}
\value{
A dataframe containing student program information.
}
\description{
Get UWL student program data
currently from WINGS report UWL_SR_REQ_TERM_BY_COLLEGE
}
